requirements.txt
python-telegram-bot
telebot
python-dotenv
google-generativeai
google.generativeai
pyTelegramBotAPI